fix: config not init in loading

+++ mod/src/main/java/io/github/settingdust/dawncraftfixes/DawncraftFixesConfig.java
@@ -16,18 +16,18 @@ public class DawncraftFixesConfig {
         COMMON_BUILDER.comment("GREATSWORD_TWOHAND_AUTO").push("greatsword_twohand_auto");
 
         COMMON_BUILDER.push("1");
-        configDamage(COMMON_BUILDER, 0.6F);
-        configAttackSpeed(COMMON_BUILDER, 1.05F);
+        configDamage(COMMON_BUILDER, 0.6D);
+        configAttackSpeed(COMMON_BUILDER, 1.05D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.push("2");
-        configDamage(COMMON_BUILDER, 0.6F);
-        configAttackSpeed(COMMON_BUILDER, 1.05F);
+        configDamage(COMMON_BUILDER, 0.6D);
+        configAttackSpeed(COMMON_BUILDER, 1.05D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.push("3");
-        configDamage(COMMON_BUILDER, 0.6F);
-        configAttackSpeed(COMMON_BUILDER, 1.05F);
+        configDamage(COMMON_BUILDER, 0.6D);
+        configAttackSpeed(COMMON_BUILDER, 1.05D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.pop();
@@ -35,43 +35,43 @@ public class DawncraftFixesConfig {
         COMMON_BUILDER.comment("GREATSWORD_DUAL_AUTO").push("greatsword_dual_auto");
 
         COMMON_BUILDER.push("1");
-        configDamage(COMMON_BUILDER, 0.55F);
-        configDamage(COMMON_BUILDER, 0.65F, 1);
-        configAttackSpeed(COMMON_BUILDER, 1.05F);
+        configDamage(COMMON_BUILDER, 0.55D);
+        configDamage(COMMON_BUILDER, 0.65D, 1);
+        configAttackSpeed(COMMON_BUILDER, 1.05D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.push("2");
-        configDamage(COMMON_BUILDER, 0.8F);
-        configAttackSpeed(COMMON_BUILDER, 1.05F);
+        configDamage(COMMON_BUILDER, 0.8D);
+        configAttackSpeed(COMMON_BUILDER, 1.05D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.push("3");
-        configDamage(COMMON_BUILDER, 0.7F);
-        configDamage(COMMON_BUILDER, 0.7F, 1);
-        configAttackSpeed(COMMON_BUILDER, 1.05F);
+        configDamage(COMMON_BUILDER, 0.7D);
+        configDamage(COMMON_BUILDER, 0.7D, 1);
+        configAttackSpeed(COMMON_BUILDER, 1.05D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.push("4");
-        configDamage(COMMON_BUILDER, 1.4F);
-        configAttackSpeed(COMMON_BUILDER, 0.75F);
+        configDamage(COMMON_BUILDER, 1.4D);
+        configAttackSpeed(COMMON_BUILDER, 0.75D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.comment("GREATSWORD_DUAL_DASH").push("greatsword_dual_dash");
-        configDamage(COMMON_BUILDER, 0.2F);
-        configAttackSpeed(COMMON_BUILDER, 1.05F);
+        configDamage(COMMON_BUILDER, 0.2D);
+        configAttackSpeed(COMMON_BUILDER, 1.05D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.comment("GREATSWORD_DUAL_AIRSLASH").push("greatsword_dual_airslash");
-        configDamage(COMMON_BUILDER, 1.2F);
-        configAttackSpeed(COMMON_BUILDER, 1.05F);
+        configDamage(COMMON_BUILDER, 1.2D);
+        configAttackSpeed(COMMON_BUILDER, 1.05D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.comment("GREATSWORD_DUAL_EARTHQUAKE").push("greatsword_dual_earthquake");
-        configDamage(COMMON_BUILDER, 1.2F);
-        configDamage(COMMON_BUILDER, 2.4F, 1);
-        configAttackSpeed(COMMON_BUILDER, 1.05F);
+        configDamage(COMMON_BUILDER, 1.2D);
+        configDamage(COMMON_BUILDER, 2.4D, 1);
+        configAttackSpeed(COMMON_BUILDER, 1.05D);
         COMMON_BUILDER.pop();
 
         COMMON_BUILDER.pop();
@@ -80,15 +80,15 @@ public class DawncraftFixesConfig {
         ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, COMMON);
     }
 
-    private static void configDamage(ForgeConfigSpec.Builder COMMON_BUILDER, float defaultValue, int index) {
-        COMMON_BUILDER.defineInRange("damage" + index, defaultValue, 0F, 100F);
+    private static void configDamage(ForgeConfigSpec.Builder COMMON_BUILDER, double defaultValue, int index) {
+        COMMON_BUILDER.defineInRange("damage" + index, defaultValue, 0D, 100D);
     }
 
-    private static void configDamage(ForgeConfigSpec.Builder COMMON_BUILDER, float defaultValue) {
-        COMMON_BUILDER.defineInRange("damage", defaultValue, 0F, 100F);
+    private static void configDamage(ForgeConfigSpec.Builder COMMON_BUILDER, double defaultValue) {
+        COMMON_BUILDER.defineInRange("damage", defaultValue, 0D, 100D);
     }
 
-    private static void configAttackSpeed(ForgeConfigSpec.Builder COMMON_BUILDER, float defaultValue) {
-        COMMON_BUILDER.defineInRange("attackSpeed", defaultValue, 0F, 100F);
+    private static void configAttackSpeed(ForgeConfigSpec.Builder COMMON_BUILDER, double defaultValue) {
+        COMMON_BUILDER.defineInRange("attackSpeed", defaultValue, 0D, 100D);
     }
 }
